% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianTiterCalc-package.R
\docType{package}
\name{bayesianTiterCalc-package}
\alias{bayesianTiterCalc-package}
\alias{bayesianTiterCalc}
\title{The 'bayesianTiterCalc' package.}
\description{
Bayesian calculation of serum antibody titers
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.9. https://mc-stan.org
}
