% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianTiterCalc-package.R
\docType{package}
\name{bayesiantiters-package}
\alias{bayesiantiters-package}
\alias{bayesiantiters}
\title{The 'bayesiantiters' package.}
\description{
Bayesian calculation of serum neutralization titer.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.9. https://mc-stan.org
}
