% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmonotone.R
\name{check_monotone}
\alias{check_monotone}
\title{Check if data is monotone}
\usage{
check_monotone(simData)
}
\arguments{
\item{simData}{A dataframe of simulated titration data}
}
\value{
A table
}
\description{
Check if data is monotone
}
\examples{
phi_Vals <- c(); ndraws = 30; a = 8.5; b = 2.5; prior_phi <- list(n=0.75, m=16);
dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10)
simData <- sample_dose_response(ndraws, prior_phi, a, b, dilutions,
nreplicates_per_dilution=2, phi_Vals)
non_monotone <- check_monotone(simData)

}
