% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evTitrations}
\alias{evTitrations}
\title{Example of serum neutralization data.}
\format{
A data frame with 80 rows and 9 variables
\describe{
\item{sampleID}{sample ID}
\item{Age}{Age of individual}
\item{Year}{Year of sample collection}
\item{titer}{Antibody titer as calculated conventionally}
\item{rep1}{Outcome of assay in first replicate, '1' for alive, '0' for dead}
\item{rep2}{Outcome of assay in second replicate, '1' for alive, '0' for dead}
\item{outcome}{sum of rep 1 and rep2}
\item{dilutions}{serum dilution, per sample}
...
}
}
\source{
Data generated using a method described in \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8386771/}
}
\usage{
evTitrations
}
\description{
A dataset containing the output of virus neutralization at eight serum dilutions. The data contains results from 10
serum samples.
}
\keyword{datasets}
