% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit a Stan model to the simulated data}
\usage{
fit_stan(
  data,
  is_sampling = TRUE,
  prior_a = list(mu = 0, sigma = 10),
  prior_b = list(mu = 0, sigma = 10),
  prior_phi = list(mu = 0, sigma = 1000),
  a = NULL,
  b = NULL,
  include_ppc = FALSE,
  n_optimisations = 5,
  ndilutions_per_id = 20,
  ...
)
}
\arguments{
\item{data}{A data frame with columns: 'dilution', 'number_surviving', 'number_replicates', 'id' (unique identifier for each sample).}

\item{is_sampling}{A Boolean (default is TRUE) indicating whether to perform sampling (if TRUE) or optimisation (if FALSE).}

\item{prior_a}{Parameters for Cauchy prior for a (by default list(mu=0, sigma=10)).}

\item{prior_b}{Parameters for Cauchy prior for b (by default list(mu=0, sigma=10)).}

\item{prior_phi}{Parameters for Cauchy prior for phi (by default list(mu=0, sigma=1000)).}

\item{a}{Fixed value for a if not estimated (by default NULL).}

\item{b}{Fixed value for b if not estimated (by default NULL).}

\item{include_ppc}{A Boolean (default is FALSE) indicating whether to include posterior predictive simulations.}

\item{n_optimisations}{Number of optimisations to perform (note optimisation is also performed as a precursor step to sampling). By default this is set to 5.}

\item{ndilutions_per_id}{Number of dilutions to sample prob(survive) at for each sample (only relevant if inc_ppc is TRUE). By default this is set to 20.}

\item{...}{Arguments passed to 'rstan::sampling' or 'rstan::optimizing' e.g., iter, chains.}
}
\value{
An object of class 'stanfit'.
}
\description{
This fits a model of the form:
}
\details{
number_surviving ~ binomial(n_replicates, theta(dilution))

where theta(dilution) is a probability of surviving which depends on the dilution.
The functional form of this is:

theta(dilution) = 1 / (1 + exp(a + b * log(phi / dilution)))

where a, b and phi are parameters. a and b can either be estimated or specified
a priori; phi is always estimated.
}
\examples{
ndraws = 30; dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10); nreplicates_per_dilution = 2
sim_data <- sample_dose_response(ndraws, dilutions, nreplicates_per_dilution)

# fit model to data not saving posterior predictive simulations
fit <- fit_stan(sim_data, iter = 100, chains = 4, init = 'random')

}
