% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{logistic function}
\usage{
logistic(concentration, a, b)
}
\arguments{
\item{concentration}{Estimate from the model}

\item{a}{Estimate from the model}

\item{b}{Estimate from the model}
}
\value{
A numeric value
}
\description{
logistic function
}
\examples{
concentration = 0.02; a = 4; b = 0.1
logistic(concentration, a, b)

}
