% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmonotonic.R
\name{non_monotonic}
\alias{non_monotonic}
\title{non monotonic logic/boolean function}
\usage{
non_monotonic(x)
}
\arguments{
\item{x}{A numeric vector, for each sample, of the number of wells with surviving cells per dilution}
}
\value{
A boolean '1' if non-monotonic and '0' if monotonic
}
\description{
non monotonic logic/boolean function
}
\examples{
x <- c(2,2,2,2,2,0,0,0)
non_monotonic(x)

}
