% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reedmuenchTitrations}
\alias{reedmuenchTitrations}
\title{Reed and Muench titration data.}
\format{
A data frame with 9 rows and 4 variables
\describe{
\item{dilution}{serum dilution}
\item{Alive}{Number of animals surviving}
\item{Dead}{Number of animals dead}
\item{Survival}{Percentage of samples that survive at each dilution}
...
}
}
\source{
A Simple Method of Estimating Fifty Per Cent Endpoints, The American Journal of Hygiene, 1938
}
\usage{
reedmuenchTitrations
}
\description{
A dataset containing the serum titration data as published in the Reed and Muench paper, 1938.
}
\keyword{datasets}
