% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledoseresponse.R
\name{sample_dose_response}
\alias{sample_dose_response}
\title{simulate data that has sample dose response}
\usage{
sample_dose_response(ndraws, dilutions, nreplicates_per_dilution = 2)
}
\arguments{
\item{ndraws}{A numeric value for number of draws or simulations}

\item{dilutions}{Numeric value for number of dilutions per sample}

\item{nreplicates_per_dilution}{Numeric value}
}
\value{
A dataframe of simulated data
}
\description{
simulate data that has sample dose response
}
\examples{
ndraws = 30; dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10); nreplicates_per_dilution = 2
simData <- sample_dose_response(ndraws, dilutions, nreplicates_per_dilution)

}
