% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_rm.R
\name{sampling_rm}
\alias{sampling_rm}
\title{Fit a Stan model to the Reed and Muench data (reedmuenchTitrations)}
\usage{
sampling_rm(standata, ...)
}
\arguments{
\item{standata}{A dataframe for constructing Stan input data}

\item{...}{Arguments passed to 'rstan::sampling' e.g., iter, chains}
}
\value{
An object of class 'stanfit'
}
\description{
Fit a Stan model to the Reed and Muench data (reedmuenchTitrations)
}
