% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_stan.R
\name{sampling_stan}
\alias{sampling_stan}
\title{Fit a Stan model to the simulated data}
\usage{
sampling_stan(inputdata, ...)
}
\arguments{
\item{inputdata}{A dataframe for constructing Stan input data}

\item{...}{Arguments passed to 'rstan::sampling' e.g., iter, chains}
}
\value{
An object of class 'stanfit'
}
\description{
Fit a Stan model to the simulated data
}
