% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDataPlot.R
\name{simData.plot}
\alias{simData.plot}
\title{Plot simulated data}
\usage{
simData.plot(simData, n)
}
\arguments{
\item{simData}{A dataframe of simulated titration data}

\item{n}{Numeric value of number of samples/draws to plot}
}
\value{
Plot
}
\description{
Plot simulated data
}
\examples{
phi_Vals <- c(); ndraws = 30; a = 8.5; b = 2.5; prior_phi <- list(n=0.75, m=16);
dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10)
simData <- sample_dose_response(ndraws, prior_phi, a, b, dilutions, nreplicates_per_dilution=2, phi_Vals)
n = 4
simData.plot(simData, n)

}
