% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatesurvival.R
\name{simulate_cell_survival}
\alias{simulate_cell_survival}
\title{simulation functions}
\usage{
simulate_cell_survival(dilution, phi, a, b, nreplicates, is_log = FALSE)
}
\arguments{
\item{dilution}{Numeric value of {8, 16, 32, 64, 128, 256, 512, 1024}.}

\item{phi}{Model estimate.}

\item{a}{Model estimate.}

\item{b}{Model estimate.}

\item{nreplicates}{Number of replicates at each dilution (by default value of 2).}

\item{is_log}{Is logistic function on log-concentration scale? (by default False).}
}
\value{
A numeric value
}
\description{
simulation functions
}
\examples{
dilution = 16; phi = 42; a = 4; b = 1.2; nreplicates = 2
simulate_cell_survival(dilution, phi, a, b, nreplicates)

}
