% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_stan.R
\name{simulations_stan}
\alias{simulations_stan}
\title{Fit a Stan model to the simulated data}
\usage{
simulations_stan(stan_data, ...)
}
\arguments{
\item{stan_data}{List that forms Stan input data}

\item{...}{Arguments passed to 'rstan::sampling' e.g., iter, chains}
}
\value{
An object of class 'stanfit'#'
}
\description{
Fit a Stan model to the simulated data
}
