% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing_stan.R
\name{optimizing_stan}
\alias{optimizing_stan}
\title{Fit a Stan model to the simulated data}
\usage{
optimizing_stan(standata, ...)
}
\arguments{
\item{standata}{A dataframe for constructing Stan input data}

\item{...}{Arguments passed to 'rstan::sampling' e.g., iter, chains}
}
\value{
An object of class 'stanfit'
}
\description{
Fit a Stan model to the simulated data
}
\examples{
ndraws = 30; a = 8.5; b = 2.5; prior_phi <- list(lower=0.75, upper=16)
dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10)
simData <- sample_dose_response(ndraws, prior_phi, a, b, dilutions, nreplicates_per_dilution=2)

stan_data = list(N = nrow(simData), nreplicates = rep(2, nrow(simData)),
survival = simData$number_surviving, dilution = simData$dilution, nsample = max(simData$draw), sample = simData$draw,
is_log = 1)

fit <- sampling_stan(standata = stan_data, iter = 100, init = 'random')

}
