% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledoseresponse.R
\name{sample_dose_response}
\alias{sample_dose_response}
\title{simulate data that has sample dose response}
\usage{
sample_dose_response(
  ndraws,
  prior_phi,
  a,
  b,
  dilutions,
  nreplicates_per_dilution = 2,
  phi_Vals
)
}
\arguments{
\item{ndraws}{A numeric value for number of draws or simulations}

\item{prior_phi}{A list with numeric values for boundaries of the uniform distribution}

\item{a}{Numeric value estimated by model}

\item{b}{Numeric value estimated by model}

\item{dilutions}{Numeric value for number of dilutions per sample}

\item{nreplicates_per_dilution}{Numeric value}

\item{phi_Vals}{An empty numeric vector initialized in the global environment before calling the function}
}
\value{
A dataframe of simulated data
}
\description{
simulate data that has sample dose response
}
\examples{
phi_Vals <- c(); ndraws = 30; a = 8.5; b = 2.5; prior_phi <- list(n=0.75, m=16);
dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10)
simData <- sample_dose_response(ndraws, prior_phi, a, b, dilutions, nreplicates_per_dilution=2, phi_Vals)

}
