% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_stan.R
\name{sampling_stan}
\alias{sampling_stan}
\title{Fit a Stan model to the simulated data}
\usage{
sampling_stan(inputdata, ...)
}
\arguments{
\item{inputdata}{A dataframe for constructing Stan input data}

\item{...}{Arguments passed to 'rstan::sampling' e.g., iter, chains}
}
\value{
An object of class 'stanfit'
}
\description{
Fit a Stan model to the simulated data
}
\examples{
ndraws = 30; a = 8.5; b = 2.5; prior_phi <- list(n=0.75, m=16)
dilutions <- 2^c(3, 4, 5, 6, 7, 8, 9, 10)
simData <- sample_dose_response(ndraws, prior_phi, a, b, dilutions, nreplicates_per_dilution=2)

fit <- sampling_stan(inputdata = simData, iter = 100, chains = 1, init = 'random')

}
